backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
anthropic==0.40.0
midiutil==1.2.1
python-dotenv==1.0.0
